Try to sync format specifiers with types being printed. In particular, avoid using "%d" to print UINT32 values as this can result in LBAs or lengths > 2147483647 being printed as negative numbers.
As part of this, drop support for antique "platforms" such as DOS where
'int' is smaller than 32 bits. This eliminates having to worry about
needing "%lu" on some platforms and "%u" on others.



The code defining types having specific numbers of bits is approaching
20 years old. We should consider switching to modern, portable ways of
doing the same thing.

+++ udf/tools/src/chkudf/display_dirs.c
@@ -196,7 +196,7 @@ int GetFID(struct FileIDDesc *FID, struct FE_or_EFE *fe, UINT16 part, int offset
     CheckTag((struct tag *)FID, location, TAGID_FILE_ID, 0, blocksize - 16);
     if (Error.Code == ERR_TAGLOC) {
       printf("** Wrong Tag Location. Expected %lld, Found %lld (%u)\n",
-         Error.Expected, Error.Found, (unsigned int) location);
+         Error.Expected, Error.Found, location);
       FID_Loc_Wrong++;
       Error.Code = 0;
     }
